Add ingestAssistant script.
In gen2, some calibrations are assumed to be curated.  Now that we are
able to construct many of these calibrations directly from the stack,
this isn't always the case.

ingestAssistant.py conceptually does the opposite of
ingestCuratedCalibs.py: It takes a calibration-like dataset from
butler, and then "exports" it into fixed path and filename.  With this
done, ingestCuratedCalibs.py can find that file, and reingest it into
butler as a calibration.

This will likely be deprecated once gen3 usage is standard.